requirements.txt
streamlit>=1.32.0
opencv-python>=4.9.0.80
numpy>=1.26.4
Pillow>=10.4.0
deepface>=0.0.93
tensorflow>=2.19.0 